move animation to its own file
