+++ src/lib.rs
@@ -1,4 +1,5 @@
 #[macro_use]
 pub mod websys_utils;
+pub mod animation;
 pub mod frames_per_second;
 pub mod universe;
